tage kinda works

+++ branch_predictor/traces.cpp
@@ -2,11 +2,11 @@
 #include "primitives.h"
 
 std::vector<BranchInstr> repeating_pattern(
-    size_t num_taken, 
-    size_t num_not_taken, 
-    size_t reps, 
-    size_t pc,
-    size_t id
+    size_t num_taken,
+    size_t num_not_taken,
+    size_t reps,
+    uint32_t pc,
+    uint32_t id
 ) {
     size_t seq_len = reps * (num_taken + num_not_taken);
     std::vector<BranchInstr> out; out.reserve(seq_len);
@@ -24,11 +24,11 @@ std::vector<BranchInstr> repeating_pattern(
 }
 
 std::vector<BranchInstr> correlated_branch(
-    size_t reps, 
-    size_t pc1,
-    size_t id1,
-    size_t pc2,
-    size_t id2,
+    size_t reps,
+    uint32_t pc1,
+    uint32_t id1,
+    uint32_t pc2,
+    uint32_t id2,
     bool anti
 ) {
     std::vector<BranchInstr> out; out.reserve(2 * reps);
@@ -45,13 +45,13 @@ std::vector<BranchInstr> correlated_branch(
 }
 
 std::vector<BranchInstr> xor_correlated_branch(
-    size_t reps, 
-    size_t pc1,
-    size_t id1,
-    size_t pc2, 
-    size_t id2,
-    size_t pc3,
-    size_t id3
+    size_t reps,
+    uint32_t pc1,
+    uint32_t id1,
+    uint32_t pc2,
+    uint32_t id2,
+    uint32_t pc3,
+    uint32_t id3
 ) {
     std::vector<BranchInstr> out; out.reserve(3 * reps);
     uint64_t state = 0x9e3779b97f4a7c15ULL;
